+++ src/services/index.js
@@ -5,3 +5,4 @@
 
 export * from './matcher/index.js';
 export * from './api/index.js';
+export * from './cache/index.js';